CORE ==== feature - data.py : added year functionality for adjust_time() feature - data.py : added timsort() function
FRAMEWORK
=========
bugfix - analysis.py : one needs to be careful, when working with climatologies, as it may happen that the climatology does NOT start in January when processed using cdos. This was e.g. the case for CERES data. The reason is, that the CDOs use always the month of the first dataset to start with. To avoid that problem, it is recommended to use timsort() function. Please have a look in the SIS analysis for more details.


git-svn-id: https://svn.zmaw.de/svn/pycmbs/trunk@95 46f1e1ae-ddaa-4461-89cc-888b818d8ef3