CORE ==== feature - data.py : added year functionality for adjust_time() feature - data.py : added timsort() function bugfix  - data.py : get_climatology() did not take into account masked areas properly. This was fixed bugfix  - data.py : get_climatology() did not calculate properly the timestamp feature - data.py : _apply_mask() now allows the mask to be also a Data object
FRAMEWORK
=========
bugfix - analysis.py : one needs to be careful, when working with climatologies, as it may happen that the climatology does NOT start in January when processed using cdos. This was e.g. the case for CERES data. The reason is, that the CDOs use always the month of the first dataset to start with. To avoid that problem, it is recommended to use timsort() function. Please have a look in the SIS analysis for more details.
feature - analysis.py : complemented albedo analysis for CERES data

git-svn-id: https://svn.zmaw.de/svn/pycmbs/trunk@96 46f1e1ae-ddaa-4461-89cc-888b818d8ef3