CORE ==== data.py - modified way how data is saved to netCDF file; now also just storing lat/lon is possible
FRAMEWORK
=========
feature - name of analysis scripts is now specified in external file in the configuration directory
feature - name of routines that specify names of functions to extract data is now specified in external file in the configuration directory
bug     - Antarctica was never used in the current setup. Now the logic is different. Antarctica is used, if the valid_mask option is 'global' or 'ocean'
          As masking Antarctica means, removing everythin below 60°S, this was not a proper approach as parts of the
          Southern Ocean were therefore masked as well !!!!

git-svn-id: https://svn.zmaw.de/svn/pycmbs/trunk@408 46f1e1ae-ddaa-4461-89cc-888b818d8ef3